Note timeout TODO, move test mocks, fix tests, change query-settings to POST
