+++ test/domain-shared/print-cost.util.test.js
@@ -1,7 +1,4 @@
-const {
-  getPrintCostNumeric,
-  noCostSettingsMessage
-} = require("../../server_src/utils/print-cost.util");
+const { getPrintCostNumeric } = require("../../server_src/utils/print-cost.util");
 
 const defaultCostSettings = {
   powerConsumption: 0.05,
